1. Remade the Unity default controllers so they're actually readable
2. Added a bunch of basic movement maneuvers:
a. press left shift to dodge in 8 directions while standing on the ground
b. press left shift in the air to boost in 8 directions in exchange for one jump
c. press space to jump vertically higher up to the max amount of air jumps

3. Player rotates (jagged and snappily) to where the camera is facing while moving
(If player is not moving, camera rotates around freely so they can look at the model or whatever)

4. Removed the stupid default model and replaced with a cylinder. The cube attached to the back of it represents Andrial's wings as a placeholder